Add files via upload
The application is a simulation of a toy robotic Arm moving on a square tabletop, of dimensions 5 units x 5 units. There are no other obstructions on the table surface. The robot is free to roam around the surface of the table, but it is prevented from falling to destruction. Any movement that would result in the robot falling from the table is prevented, however further valid movement commands is still allowed.